perf: avoid double serialization
If you do not pass a `CID` instance to `ipld.put` right now it does
the following:

1. get the `format` of a node
2. get the `CID` of a node using `format.util.cid`
3. get the serialized form using `format.util.serialize`
4. write the serialized form into the blockstore using the `CID`

The way these formats typically calculate the `CID` of a node is by
serializing it first, so during the above operations a node is
serialized twice.

According to the [spec](https://github.com/ipld/interface-ipld-format#utilcidbinaryblob-options-callback)
each format's `format.util.cid` method  should accept a binary blob
as an argument (I read that as a `Buffer`).

This PR changes the order above to serialize first, then pass
the serialized form to the relevant format implementation as per
the spec.

It seems the only format implementation that accepts a binary blob
is `dag-pb` so some extra work will need to be done to the other
formats before it can be merged.