feat: return cid of last node traversed from ipld.get and friends
We can pass a path to `ipld.get` that makes the resolver attempt to
do a graph traversal until it runs out of path segments to resolve.

At that point we don't know the `cid` of the node that will be returned,
which things like the `unixfs-exporter` need to know.