refactor: remove inMemory util
This removes the dependency on `ipfs-repo`, `ipfs-block-service` and `interface-datastore` and resolves #151.

BREAKING CHANGE: This module no longer exports an `inMemory` utility to create an IPLD instance that uses a block service that stores data in memory. Please use the [`ipld-in-memory`](https://www.npmjs.com/package/ipld-in-memory) module instead.

License: MIT
Signed-off-by: Alan Shaw <alan.shaw@protocol.ai>