refactor: store codecs by their code
Instead of storing a codec by their name, use their code instead.

The tests are changed from `base1` to `blake2b-8` as codecs for different
bases are no longer part of multicodec, but are part of multibase now.

BREAKING CHANGE: The `codec` parameter in `options.loadFormat()` is a number

Instead of returnign the name of the codec as string, the codec code (a number)
is now returned.

So if you e.g. check within the function for a certain format, it changes from:

    async loadFormat (codec) {
      if (codec !== 'dag-cbor') …
    }

To:

    async loadFormat (codec) {
      if (codec !== multicodec.DAG_CBOR) …
    }