refactor: make code promisify free
There's no need for the promisify-es6 module anymore as all the code
is using async/await based libraries now.

This is the same commit as 1a98e165e4f3351d2226b885241bc3615ded9763,
the only difference is a commit message that contains the information
about the breaking change.

This reverts commit e84955b12fe51a171ff771a39d9dae28fec88b6c.

BREAKING CHANGE: Everyone upgrading to this release also needs to upgrade
ipfs-block-service, ipfs-repo and ipld-in-memory.

This commit uses the new async API of:

 - ipfs-block-service >= 0.16.0
 - ipfs-repo >=0.27.0
 - ipld-in-memory >= 3.0.0

If your library has a dependency on any of those, you likely also need
to upgrade to at least those versions.