feat: expose getFormat function
It's useful to be able to get the format implementations out of ipld
just before `ipld.put` operations in the HTTP API server of IPFS.

This is because we've serialized the node on the client for transmission
over HTTP so need to deserialize it before passing it into `ipld.put`
for it to be serialized (and as such, verified) again.

We can duplicate the list of resolvers and logic behind the addition
of new resolvers from ipld or we can just expose the `getFormat` function
which we do here to reduce any code duplication.